removed unused methods from bookmarkClient. renamed to bookmarkService
