+++ src/client-app/components/bookmark/BookmarkHome.jsx
@@ -1,5 +1,5 @@
 import React, { Component, PropTypes as T} from 'react'
-import {getRecentBookmarks} from '../../common/bookmarkClient'
+import {getRecentBookmarks} from '../../common/bookmarkService'
 import BookmarkSearch from './BookmarkSearch'
 import BookmarkRow from './BookmarkRow'
 import {Link} from 'react-router'
